+++ packages/adapters/uni-uploadFile/index.ts
@@ -1,5 +1,6 @@
 import { Adapter } from '@applet-request/core';
 import type { RequestContext, MiddlewareNext, RequestConfig } from '@applet-request/core';
+import { BaseException } from '@applet-request/shared';
 
 /**
  * uni.uploadFile的其他请求配置
@@ -10,6 +11,12 @@ export type UniUploadFileOtherConfig = Omit<UniApp.UploadFileOption, 'url' | 'fa
  * uni.uploadFile的请求配置
  */
 export type UniUploadFileConfig = RequestConfig<UniUploadFileOtherConfig>;
+
+export class UniUploadFileFailException extends BaseException<UniApp.GeneralCallbackResult> {
+  constructor(message: string, raw?: UniApp.GeneralCallbackResult, options?: ErrorOptions | undefined) {
+    super(message, raw, options);
+  }
+}
 
 /**
  * uni.uploadFile的请求适配器
@@ -39,7 +46,7 @@ export class UniUploadFileAdaptor<Data> extends Adapter<
           resolve(res);
         },
         fail(err) {
-          reject(err);
+          reject(new UniUploadFileFailException(err.errMsg, err));
         },
       });
     });