+++ packages/core/types/exceptions.ts
@@ -1,0 +1,10 @@
+export enum BaseExceptionMessage {
+  /**
+   * 响应体data为空
+   */
+  RESPONSE_DATA_IS_EMPTY = 'response data is empty',
+  /**
+   * 请求失败（默认兜底异常）
+   */
+  REQUEST_FAIL = 'request fail',
+}