+++ packages/core/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @applet-request/core
 
+## 0.1.11
+
+### Patch Changes
+
+- [`18eb287`](https://github.com/calandnong/applet-request/commit/18eb2870faad51c30c6754f601080d7b1e2034d8) Thanks [@calandnong](https://github.com/calandnong)! - feat: 底层增加异常兜底机制
+
+- [`4e396ca`](https://github.com/calandnong/applet-request/commit/4e396cafe58f8c27a1e1b0e2c07a55ccd7b61c39) Thanks [@calandnong](https://github.com/calandnong)! - fix: fix merge formdata
+
+- [`18eb287`](https://github.com/calandnong/applet-request/commit/18eb2870faad51c30c6754f601080d7b1e2034d8) Thanks [@calandnong](https://github.com/calandnong)! - feat: 发布正式版
+
+- feat: 发布正式版
+
+- [`8518954`](https://github.com/calandnong/applet-request/commit/85189545d98c56887bb01cc4e256be8b9ffd2cd9) Thanks [@calandnong](https://github.com/calandnong)! - fix: 修复 FormData 判断问题
+
+## 0.1.11-beta.2
+
+### Patch Changes
+
+- feat: 底层增加异常兜底机制
+
+## 0.1.11-beta.1
+
+### Patch Changes
+
+- fix: 修复 FormData 判断问题
+
+## 0.1.11-beta.0
+
+### Patch Changes
+
+- fix: fix merge formdata
+
 ## 0.1.10
 
 ### Patch Changes